Refactor magician components and update imports for spells and spirits
