+++ src/components/runner/magician/magicianUtils.ts
@@ -1,7 +1,5 @@
 import { AwakeningType } from "#/system/awakeningType.ts"
 
-export const SpellsBpPerSpell = 3
-
 export const isMagician = (awakeningType: AwakeningType) => {
   return (
     awakeningType === AwakeningType.Magician
