+++ src/components/runner/magician/spells/dialogs/spellFormDialog.tsx
@@ -3,8 +3,8 @@ import Button from "@mui/material/Button"
 import Stack from "@mui/material/Stack"
 import type { FC } from "react"
 
-import { SpellFormFields } from "#/components/runner/spells/form/spellFormFields.tsx"
-import { useSpellForm } from "#/components/runner/spells/form/useSpellForm.ts"
+import { SpellFormFields } from "#/components/runner/magician/spells/form/spellFormFields.tsx"
+import { useSpellForm } from "#/components/runner/magician/spells/form/useSpellForm.ts"
 import type { ControlledDialogProps } from "#/components/ui/dialog/controlledDialogProps.ts"
 import { ControlledDialog, Dialog } from "#/components/ui/dialog/dialog.tsx"
 import { useDialog } from "#/components/ui/dialog/useDialog.tsx"
