+++ src/components/runner/magician/spells/drainResistanceDicePool.tsx
