+++ src/components/runner/magician/spells/drainValue.tsx
