+++ src/components/runner/magician/spells/spellCastSection.tsx
