+++ src/components/runner/magician/spells/spellViewerListItem.tsx
