+++ src/components/runner/magician/spells/spellsUtils.ts
@@ -1,0 +1,4 @@
+import { BuilderConfig } from "#/components/builder/builderConfig.ts"
+
+/** @deprecated Use `BuilderConfig.magic.spells.bpCost` instead. */
+export const SpellsBpPerSpell = BuilderConfig.magic.spells.bpCost
