+++ src/components/runner/magician/spells/spellsViewerSection.test.tsx
