+++ src/components/runner/magician/spells/spellsViewerSection.tsx
