+++ src/components/runner/magician/spirits/critterPowerChip.tsx
