+++ src/components/runner/magician/spirits/dialogs/spiritFormDialog.tsx
@@ -3,9 +3,9 @@ import Divider from "@mui/material/Divider"
 import Stack from "@mui/material/Stack"
 import type { FC } from "react"
 
-import { SpiritFormFields } from "#/components/runner/spirits/form/spiritFormFields.tsx"
-import { useSpiritForm } from "#/components/runner/spirits/form/useSpiritForm.ts"
-import { SummoningSection } from "#/components/runner/spirits/summoningSection.tsx"
+import { SpiritFormFields } from "#/components/runner/magician/spirits/form/spiritFormFields.tsx"
+import { useSpiritForm } from "#/components/runner/magician/spirits/form/useSpiritForm.ts"
+import { SummoningSection } from "#/components/runner/magician/spirits/summoningSection.tsx"
 import { Dialog } from "#/components/ui/dialog/dialog.tsx"
 import { useRunnerStoreSelector } from "#/stores/runner/runnerStore.selectors.ts"
 import type { SpiritData, SpiritType } from "#/system/magic/spiritData.ts"
