+++ src/components/runner/magician/spirits/form/spiritFormFields.tsx
