+++ src/components/runner/magician/spirits/spiritConditionMonitor.tsx
