+++ src/components/runner/magician/spirits/spiritItemCard.tsx
