+++ src/components/runner/magician/spirits/spiritList.test.tsx
