+++ src/routes/$runnerId/spells.tsx
@@ -3,7 +3,7 @@ import Stack from "@mui/material/Stack"
 import Typography from "@mui/material/Typography"
 import { createFileRoute } from "@tanstack/react-router"
 
-import { SpellsViewerSection } from "#/components/runner/spells/spellsViewerSection.tsx"
+import { SpellsViewerSection } from "#/components/runner/magician/spells/spellsViewerSection.tsx"
 import { useRunnerStoreSelector } from "#/stores/runner/runnerStore.selectors.ts"
 import { AwakeningType } from "#/system/awakeningType.ts"
