+++ src/routes/$runnerId/spirits.tsx
@@ -1,7 +1,8 @@
 import Stack from "@mui/material/Stack"
 import { createFileRoute } from "@tanstack/react-router"
 
-import { SpiritList } from "#/components/runner/spirits/spiritList.tsx"
+import { SpiritList } from "#/components/runner/magician/spirits/spiritList.tsx"
+import { TraditionDisplay } from "#/components/runner/magician/spirits/traditionDisplay.tsx"
 import { SectionHeader } from "#/components/ui/text/sectionHeader.tsx"
 
 export const Route = createFileRoute("/$runnerId/spirits")({
@@ -12,6 +13,8 @@ function RouteComponent() {
   return (
     <Stack sx={{ gap: 1 }}>
       <SectionHeader>Spirits</SectionHeader>
+
+      <TraditionDisplay />
 
       <SpiritList />
     </Stack>
